Update 07_sequence_tracking.R
Two lines of code, using 'file.exists', added to ensure only rows/samples in the 'out' object and sample names in the 'sample.names' object for which 'filtFs' has corresponding files in the filterAndTrim directory are retained. This should remove the misalignment of rows and duplicate number entries in the 07_track_reads_table.tsv.